requirements.txt
# RAGTrace - Multi-LLM RAG System Performance Evaluation Platform
# Generated from pyproject.toml - use `uv pip install -r requirements.txt` to install

# Core Evaluation Framework
dependency-injector>=4.48.1
ragas>=0.1.0
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0

# Web Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# HTTP Requests and Network
requests>=2.31.0

# Performance Monitoring (Optional)
# watchdog>=3.0.0  # For better Streamlit performance - use `uv pip install ragtrace[performance]`